+++ R/gridcv.R
@@ -14,10 +14,19 @@ gridcv <- function(X, Y, segm, score, fun, pars, verb = TRUE) {
             if(verb)
                 cat("segm=", j, " ", sep = "")
             s <- sort(listsegm[[j]])
-            zres[[j]] <- gridscore(
+            if(is.list(X)){
+              zres[[j]] <- gridscore(
+                lapply(1:length(X), function(k) X[[k]][-s, , drop = FALSE]), 
+                Y[-s, , drop = FALSE],
+                lapply(1:length(X), function(k) X[[k]][s, , drop = FALSE]), 
+                Y[s, , drop = FALSE],
+                score = score, fun = fun, pars = pars)
+            }else{
+              zres[[j]] <- gridscore(
                 X[-s, , drop = FALSE], Y[-s, , drop = FALSE],
                 X[s, , drop = FALSE], Y[s, , drop = FALSE],
                 score = score, fun = fun, pars = pars)
+            }
         }
         zres <- setDF(rbindlist(zres))
         res_rep[[i]] <- cbind(rep = rep(i, nsegm * nco),
@@ -50,10 +59,19 @@ gridcvlv <- function(X, Y, segm, score, fun, nlv, pars = NULL, verb = TRUE) {
             if(verb)
                 cat("segm=", j, " ", sep = "")
             s <- sort(listsegm[[j]])
-            zres[[j]] <- gridscorelv(
+            if(is.list(X)){
+              zres[[j]] <- gridscorelv(
+                lapply(1:length(X), function(k) X[[k]][-s, , drop = FALSE]), 
+                Y[-s, , drop = FALSE],
+                lapply(1:length(X), function(k) X[[k]][s, , drop = FALSE]), 
+                Y[s, , drop = FALSE],
+                score = score, fun = fun, nlv = nlv, pars = pars)
+            }else{
+              zres[[j]] <- gridscorelv(
                 X[-s, , drop = FALSE], Y[-s, , drop = FALSE],
                 X[s, , drop = FALSE], Y[s, , drop = FALSE],
                 score = score, fun = fun, nlv = nlv, pars = pars)
+            }
         }
         zres <- setDF(rbindlist(zres))
         ## Case where pars is empty
@@ -96,10 +114,19 @@ gridcvlb <- function(X, Y, segm, score, fun, lb, pars = NULL, verb = TRUE) {
             if(verb)
                 cat("segm=", j, " ", sep = "")
             s <- sort(listsegm[[j]])
-            zres[[j]] <- gridscorelb(
+            if(is.list(X)){
+              zres[[j]] <- gridscorelb(
+                lapply(1:length(X), function(k) X[[k]][-s, , drop = FALSE]), 
+                Y[-s, , drop = FALSE],
+                lapply(1:length(X), function(k) X[[k]][s, , drop = FALSE]), 
+                Y[s, , drop = FALSE],
+                score = score, fun = fun, lb = lb, pars = pars)
+            }else{
+              zres[[j]] <- gridscorelb(
                 X[-s, , drop = FALSE], Y[-s, , drop = FALSE],
                 X[s, , drop = FALSE], Y[s, , drop = FALSE],
                 score = score, fun = fun, lb = lb, pars = pars)
+            }
         }
         zres <- setDF(rbindlist(zres))
         ## Case where pars is empty